requirements.txt
pypng
pyqrcode
pillow
qrcode
flask